Allow display of plantUML graph and matrix from selecting graphical elements
The functionality is now generic and depends on the handlers to get the selected Object as an EObject.
The code also contains functionality that allows to use toggle transitivity to switch between seeing the selected elements only and seeing the selected elements and their children in the Matrix view. Test cases for this also added.

Change-Id: Ia873a48665476bfa4cf906877f5d0e58a53b5cc1