Selection of the right priority handler now based on type hierarchy
If several artifact handlers are eligible, the most specific one is
chosen. The specificity of a handler is determined by which of the
classes handled is more specific. For instance, if the EmfHandler and
the BuildElementHandler are available to handle the element, the
BuildElementHandler will be selected since BuildElement implements the
EObject interface and EmfHandler can only handle EObjects.

Change-Id: If8d77c237b64ba55a02ca9b0d386d02b9790f2b5