Added ability to show graph only for selected elements.
If several model elements are selected at the same time, Capra can now
show a graph instead of a matrix with only these elements. This
behaviour can be (de)activated via a menu entry.

This commit also changes the behaviour of the menu entries and makes
them toggleable to better conform to user expectations.