Made notification feature generic
The existing notification feature was dependent on a specific trace metamodel.
This commit makes the notification feature independent of the trace metamodel by adding two methods to the TraceMetaModelAdapter interface.
A method to get all trace links from a trace model and a method to delete specific links from a trace model.

Change-Id: I956a74357680e9c441d0fabccd8144e12fa39f76