Releng: added target definitions for Oxygen and 2018-09
We are now using the Eclipse CBI TPD infrastructure to generate
target definitions for the different Eclipse platforms. That means
that the releng.target project has been converted to an xtext
project to enable the generation of the target definition files.
For now, the target definition we use for building will remain the
hand-crafted one we used so far, but in the future, we might move
to one with dependencies to a newer Eclipse platform. Developers
can already start to adapt Eclipse Capra to newer versions, though.
While Oxygen works without issue, there are dependency issues in
the RegIF Handler and in the Zest visualisation for 2018-09..