PlantUML view now allows navigating to artifacts
The PlantUML view now displays links to the artifacts that
are either resolved from the ArtifactWrapper or from the
EObject. On clicking the link, PlantUML opens the default
editor associated with the referenced resource. This works
well for all code editors, ProR, UML, and others. However,
it requires that the URI is stored in the ArtifactWrapper
correctly, a requirement that is not yet true for all
ArtifactHandlers (e.g., for FeatureIDE).

Change-Id: I420e899d60dfb886bf5e8ff0fc1e076a618bf822