Improved handling of artifact URIs in some handlers
The changes in commit c41d6cca make it necessary that each artifact
can be identified with its URI. Most handlers have already stored a
correct URI relative to the workspace in the ArtifactWrapper, but
some have not been quite as disciplined. This commit changes this
behaviour so that all ArtifactHandlers now store correct URIs for
the artifacts in the corresponding wrappers. This will allow links
to almost all artifacts in the PlantUML view.