Bug 517292 - [Feature] Prevent the creation of trace links that already exists
This is done by first creating a Connection that represents a trace link to be
created and comparing this to existing connections in the trace model.
In the current implementation, Connection A is equal to Connection B
if it contains the same elements (origin and targets) irrespective of
the order and has the same trace link type.

Change-Id: Iefcc574a8b19138828e020ba800c660176bbc640