Bug 512417: Traceability visualisation updated when selection changes
So far, the PlantUML visualisation of the traceability model only
updated if the selection within an editor was changed. This was due to a
limitation of PlantUML which only picked up on selection changes in
editors. Starting with PlantUML 1.1.23, this is no longer the case and
all views in Eclipse now trigger an update of the PlantUML view.

This patch ensures that the new functionality is exposed in Eclipse
Capra. It sets the minimum version of PlantUML to use to 1.1.23 and
makes the necessary addition to the UI and the testsuite. It also makes
the visualisation more robust by capturing some situations in which NPEs
could occur previously.

Change-Id: I4f353ff63b06dc867228640a0b4c4519f9c2b6ea