UI bundles now use icons from main ui bundle when possible
Each UI bundle used to have it's own copy of shared icons. With this
patch, those bundles that reference org.eclipse.capra.ui use the icons
from this bundle. This avoids duplicating the files in the different
bundles. The notable exception is org.eclipse.capra.ui.office which does
not reference org.eclipse.capra.ui and thus has it's own set of icons.