[EPLv2] Updated features to reference EPL v2
This commit adds the EPL v2 license to all of Capra's features.
It also adds a Software User Agreement and an about page as per the
standards of the Eclipse foundation. In addition, the strings from
feature.xml are externalised, and the descriptions and the copyright
message have been extended and improved.

This commit is one part of addressing bug 546480.