Bug 527440 - Make tests run reliably on build server
The C tests are always hanging with an error that says
"The user operation is waiting for Discover compiler built-in
language to complete." We are not sure why this randomly occurs,
but it seems to be connected to a race condition since it mostly
happens under high load. This commit is an attempt to fix this
issue by using an ICProject directly instead of IProject.