[EPLv2] Update copyright messages in all relevant files
This is the second commit to ensure that we conform to EPLv2. In this
commit, the copyright messages of all relevant files have been changed
to reflect the new license. At the same time, the copyright messages
now include the current revision year, an SPDX-License-Identifier,
and list Chalmers | University of Gothenburg as contributors of
additional features and new API.

This commit addresses Bug 546480.