Bug 546829: Increased waiting time for notification tests
The notification tests test if the respective markers are created
when artifacts connected by trace links change. The creation of
markers take some time and therefore a UI_REACTION_WAITING_TIME
was set to 500 milliseconds. This seems not enough as the tests still
fail sometimes when the server is busy. To deal with this, the waiting
time has been increases to 1000 milliseconds.