[Releng] Added Oomph setup file
The Oomph setup file contained in this commit allows
developers to install an Eclipse with all necessary tools
via the Eclipse Installer. This should significantly simplify
getting started with contributions to Eclipse Capra. The
setup file is based used together with the "Eclipse IDE for
Eclipse Committers" product offered by the Eclipse Installer.
Apart from installing all necessary plugins, the project setup
will also download the code, activate the correct target
platform, and connect the developer to Gerrit and Bugzilla.

Depending on which Eclipse version is chosen, the CBI
Targetplatform tools will be installed as well.