Bug 547077: Upack packed items for display in plantUML viewer
When an element that is packed in a container is selected,
the plantUML viewer tries to display its diagram which ends up
with an exception because the handler exists for the element contained
in the container and not the actual container. To solve this, the unpack
method is called for every handler that implements the IArtifactUnpacker
interface.