Bug 517159: Use meaningful display name for spreadsheets
Updated getDisplayName() for the office handler to use the
content in the column selected by the user in the office
preference page.

Also updated the createWrapper() to store the data of
the row and not only the id. This makes sure that resolveWrapper()
has enough data to recreate the row with its contents.