Show confirmation after a trace link was created
When a trace a link has been created successfully, an information
dialog with the message "Trace link has been successfully created"
will show. When the user selects "Do not show this dialog again"
the setting is stored in Capra preferences and the dialog will
not show up again when a trace link is created.

Change-Id: Ic58faa92362442ef5baa9911350120da25fecf46