Refactored trace creation into an undoable operation
The creation of a trace link is now an undoable operation. This means
that all views that want to create a trace link can reuse the operation
from the org.eclipse.capra.ui.operations package and get automatic undo
ability. To facilitate this, a new getTraces() method has also been
implemented in TraceHelper. In addition, the API in this class has been
simplified.
At the same time, creating a trace link with the SelectionView can now
be undone via the Edit menu as well as using the platform-specific keys.

Change-Id: I5042ba23fce8eccceca3ebe8fdfa11080ff0b8c7