Added ability to select trace type in SelectionView
The selection view has been extended with a combo box that contains the
applicable trace type for any set of artifacts that is dropped into the
view. The combo box is updated whenever the selection changes. If the
user selects an applicable trace type from the combo box, the link is
created immediately. If no trace type is selected, a dialog for
selecting it pops up as before.

This feature has been inspired by the AMASS version of Eclipse Capra.