Added a toolbar to the Capra Selection View
A toolbar has been added that offers buttons for trace creation and for
clearing the selection. This patch also cleans up the plugin.xml file
significantly and introduces consistent naming conventions for all
commands, views, etc. A category for commands has been added as well to
make it easier to assign keys to the different Eclipse Capra commands.

Change-Id: Ie96ead24584da2afce17148d46dbdff1cce35e5a