[fix] Ensured that selection views commands are enabled
A previous commit changed the names of the commands in the main UI
bundle. These changes have not been made everywhere, causing the Capra
context menu as well as the toolbar buttons to be grayed out. This
prevented the creation of new trace links and adding links to the
selection. This patch re-enables these commands by using correct command
ids everywhere.