Harmonised bundle names and vendors
So far, we were not following the Eclipse standard conventions for
naming bundles and for providing a bundle vendor. This change fixes
that. All bundles are now prefixed with Eclipse Capra and the bundle
vendor is Eclipse Capra as well. At the same time, these strings have
been externalised into a plugin.properties or bundle.properties file.

Change-Id: I70fd8bc84675e52bb11bcfabe1086cfdb0aa468a