Added target definition file for Eclipse 2019-06
While the preferred method of setting up the Eclipse Capra development
environment is to use our Oomph setup, it is still helpful to have
target definition files for the different Eclipse version. This commit
adds target definitions for the current 2019-06 edition and fixes a typo
in the target definitions for 2019-03.