Improve/fix EclipseCapra.setup
Fix the Eclipse Ini task for oomph.redirection.capra to specify a URI
without a query string on the left and to specify the proper location
within the clone of the actual setup file.  This way, the setup file in
the clone is use once the clone exists so that you (or I) can make
changes to it and test them before committing them.

Also provide support for Git and Gerrit URI choices so that anyone can
choose to set up an IDE where they can commit Gerrit changes, as I have
done in this case.

Change-Id: Ib1c8a1dad47084f9c26e53e2f6170af7b605f90f
Signed-off-by: Ed Merks <ed.merks@gmail.com>