Updated project meta-data
This change addresses several issues with the project meta-data. In
particular, all exported packages of the bundles are now explicitly
declared, the SLF4J fragment is included to avoid warnings about missing
logger classes when running the tests and Maven, and the OSGI folders
are now explicitly included in the binaries. While none of these issues
were critical, they created unnecessary warnings in Eclipse or in Maven.