Moved helper methods for artifacts into o.e.c.core package
Two handy utility functions have been hidden inside a class within the
PlantUML visualisation. One function, getArtifactLabel(), retrieves a
label for an EObject, regardless of whether it is an ArtifactWrapper or
an object we can deal with without an ArtifactHandler. The other
function, getArtifactLocation(), does the same thing with the URI of the
object and also ensures that we get a platform URI. Since these
functions are also helpful in other visualisations, they should be
accessible through a core class. Now, they reside in the ArtifactHelper
which allows us to use them anywhere in Eclipse Capra.

Change-Id: I3bc3bac6d16e6aeeae46585049a32088e68f8ad0