Fixed calculation of transitivity depth for generic trace meta model
The visualisations did not display the correct depth of transitive links
since the calculation of how deep to go depended on a constant that was
set to -2. That meant that there were always at least three levels of
transitive links explored. This commit changes this default to 1 which
produces the expected behaviour in the visualisation.