[Fix] Fixed issue with deleteTraces() not removing trace links
The previous implementation of GenericMetaModelAdapter.deleteTraces()
compared the object references of the traces to delete with the traces
in the trace model. This meant that in many situations, traces that
should have been deleted were not removed. Now, EcoreUtils.equals() is
used instead to compare the trace links structurally. In addition,
TestCreateTraceOperation has been extended and now creates two links, of
which one operation is undone. The test then checks for the absence of
the one whose creation was undone and the presence of the other.