Bug 550301: Fixed issue with DeleteQickFix not deleting the trace link(s)
The delete quick fix did not work because the correct trace
link to be deleted could not be found using the current
getConnectedElements() method in the GenericMetamodelAdapter class.

This method is now updated to use the EcoreUtil.equals method
to check if an artifact is part of a trace link.
The change also ensures that the right artifact wrapper is deleted.

If the deleted element is an EObject, the EMF editor deletes the element
from the trace link directly. Therefore, the code checks if there is
a trace link with only one item and deletes the trace link. If the
existing link has more than one item, it is not deleted.

Change-Id: Ic79840013f243d252382847f2ad99466d94117c8