Added selection support for Xtext editors and the Xtext outline view
Xtext internally uses EObjects to represent elements of the language.
This means that Eclipse Capra natively supports Xtext. However, it was
not possible to move elements from the Xtext editors or the Xtext
outline views to the Capra selection view since these workbench parts
use a different internal representation. This change adds support for
unpacking this representation and thus allows creating trace links from
and to Xtext elements directly from the editor and outline view.

Signed-off-by: Dominique Blouin <dominique.blouin@laposte.net>
Change-Id: I45fb2c08f3e07dd7044620ec5113cbe7f3e62141