Added extension point to support for unpacking selections from parts.
This change is made so that elements Capra can create traces to can be
extracted from ISelection instances that contain complex objects. Via
the new ISelectionSupport interface and the corresponding extension
points, plugins can register their own SelectionSupport clases that
take over the unwrapping.
SelectionSupportHelper will take over the functionality that has so
far been a part of TraceCreationHelper.

One example of this is the upcoming handler for Xtext. Xtext workbench
parts such as the Xtext editor and the outline part use selections that
contain instances of specialised Xtext classes. This means that  the
Xtext API must be used to extract the elements from the selection. In
order to avoid introducing dependencies to the Xtext framework into
the core Capra, an extension of the selection support extension point
will be used.

Signed-off-by: Dominique Blouin <dominique.blouin@laposte.net>
Change-Id: Iba72d37918474b7a6e0a0f1273a198de6fe2eaae