Pushed dependency of the PlantUML Eclipse plugin to 1.1.24
Version 1.1.24 of PlantUML fixes a number of issues that are relevant
for Eclipse Capra. In particular, it adds the ability to
programmatically update the view
(https://github.com/hallvard/plantuml/issues/92). A future update is
going to use this capability to automatically update the traceability
graph or matrix whenever the view options (e.g., transitivity depth)
change.