Fixed issue with UI of Office handler not loading
The plug-ins related to the user interface of the Office
handler were not correctly loading. This was due to an
unresolvable dependency conflict since both
org.eclipse.core.runtime and org.dom4j provide
javax.xml.bind, but in different version. This quick
workaround addresses the issue by introducing an
explicit dependency to javax.xml.bind version 2.2.0
which is the one provided by org.dom4j.

Change-Id: I2e4d38b93a0a3a09317b9df1dfbbd9552c246f09