Added sunburst visualisation of trace model
This commit adds a visualisation of the trace model in the form
of a sunburst diagram. The currently selected artifact is in the
centre of the diagram and other artifacts traced to it are
represented as circle sections around the centre. The current
visualisation shows diagrams to a depth of a maximum of 5.

The implementation uses a browser view and the JavaScript library
at https://github.com/vasturiano/sunburst-chart for display. The
diagram is redrawn whenever a new artifact is selected and supports
the selection of multiple artifacts.

This work was adapted from a contribution by Fredrik Johansson and
Themistoklis Ntoukolis who made the code developed in their master's
thesis available under EPLv2.

Change-Id: I86d8c1ce5f177e2c43625d219ef2afcf7243e14f