Added helper methods to the TraceHelper class
This commit adds two helper methods, isArtifactInTraceModel(EObject artifact)
to check if an artifact has trace links and is therefore present
in the trace model and getTraces(List<EObject> artifacts)
to give a list of connections involving a given a list of
artifacts. Tests for the getTraces(List<EObject> artifacts)
methods have also been added.

Additionally, the getTracedElements(Connection connection) method
is modified to make sure that it returns a list of unique traced
artifacts.

Change-Id: Ibb128cf5a530629e79ff3ac3df96ee80af5ea59b