Added ability to select traversal depth to Sunburst view
This change introduces the ability to select how deep the trace model
should be traversed when rendering the Sunburst view. The user has the
ability to select the value from the view's pull down menu. The selected
value will be stored in the Eclipse preference store and read when the
view is loaded.

Change-Id: Idd48953a68332b39e0c0972bcd415b84c6747f7b