Added an interactive traceability matrix view using NatTable
The traceability matrix shows all selected artifacts and the artifacts
they are directly connected to via trace links. It also displays the
type of link between the artifacts, allows navigating to the artifact by
double-clicking the column header, and shows information about the
artifact and the trace link as a tool-tip. It is also possible to export
the currently viewed traceability matrix as a Microsoft Excel file.

This work was adapted from a contribution by Fredrik Johansson and
Themistoklis Ntoukolis who made the code developed in their master's
thesis available under EPLv2.

Change-Id: Ia132619593caec2e164ef96f3b1e46d557c613b9