Cleaned up features and categories for update site
This commit introduces separate features for the two new visualisation
options (sunburst and matrix). Both new features are now included in the
visualisations category of the update site. The commit also moves the
Xtend support to the handlers category. While the Xtend support is not
strictly speaking a handler, it supports working with a certain type of
artifact and is thus better suited to that category.