Added support for showing connections and artifacts in the property view
In this commit, an adapter factory and two adapters are added that
provide initial support to show artifacts and trace links in the
properties view. For trace links, all relevant generic properties
(origin, targets, type) as well as link type-specific metadata are
shown. For artifacts, location and label as well as artifact-specific
metadata are shown. This generic implementation can be used by
registering the provided factory to the o.e.c.r.adapters extension point
and implementing an ISelectionProvider in the view that returns a
Connection or an EObject.