Adds support for the properties view to the traceability matrix
Using the adapter factory and the adapters from the ui bundle, this
commit adds support for showing the properties of trace links and
artifacts in the Eclipse properties view. When a trace link is selected
in the matrix, its properties are shown, when an entire column or row
are selected, the properties of the artifact it represents are shown.

Change-Id: I275a25afefe2d557c010c9849c55e8e2933590d7