Refactored new views to use methods from core packages
This commit introduces refactorings of the sunburst and the matrix view
that make use of the new methods introduced in TraceHelper. It also adds
shared functionality from those views to EMFHelper and TraceHelper to
make it available to other bundles. In particular, a simple comparison
of EObjects based on their identifier was added to EMFHelper and a
method to get a list of the elements that are traced by a list of
connections.

Change-Id: I948d7a1cd29ec3379bd00ea65684c09723c42ac5