Fixed a bug where the properties of a connection would not show
This commit addresses a bug that occurred when a user first selected a
trace link in the matrix view to show its properties, then switched to a
different matrix and selected another link. The expected behavior is
that the properties of this new link are shown, but in practice, either
no properties or the properties of the previously selected link was
shown. The bug has been fixed by updating the SelectionProvider with the
SelectionLayer and DataProvider whenever a new traceability matrix is
shown.