[WIP] PlantUML view now updates automatically on preference change
So far, the PlantUML traceability view was only rendered anew when the
artifact was changed. This commit uses an extension of PlantUML that is
not part of the official PlantUML for Eclipse plugins to update the view
whenever one of the options in the pulldown menu has been changed.

Marking this as WIP as long as the necessary capability is not yet in
PlantUML.