PlantUML view now shows correct state of preferences in menu
The PlantUML view has a pull-down menu which allows to select several
options, including whether transitive links are shown or not. The
selected state is stored in the preferences. However, these options were
not yet visible in the toggle state of these options in the menu. With
this fix, the state of the options is now read from the preferences when
the view is initialised and the toggles are set correctly.

Change-Id: Id136063a0ca455cfae4e74f6081b693986c0e4d4