Refactored rendering code for PlantUML visualisation
This commit refactors the code that is responsible for compiling the
data for the PlantUML visualisation. The main purpose is to make the
code easier to read and maintain by reducing the number of case
distinctions and by extracting parts of the code into separate methods.
At the same time, this commit introduces a number of bug fixes that
mainly have to do with the way the preferences for the view are handled,
how the traceability matrix is rendered, and how the graph is rendered
when several artifacts are selected.