Updated PlantUML viewer user instructions
As long as no element is selected, the PlantUML viewer shows an
instruction pane that asked the user to "choose two elements". This has
been updated to also include selecting a single element and changing the
language from "choose" to "select" to make it clearer to the user how to
do this.
Change-Id: I6c71ae4d91e49931d94b093e423891ad4731dcb8