Fixed a bug where selected element was not displayed in trace viewers
In some circumstances, the selected element, e.g., in the Papyrus
diagram editor, was not displayed in the trace viewers, e.g., in the
PlantUML View. Instead, a different element that is a child of the
selected element was displayed. This was due to the fact that
EMFHelper.linearize() put the root of the element tree last in the list.
This commit fixes the behaviour. The change also reorders the elements
in the traceability matrix view to show the container first and
contained elements afterwards.
Change-Id: I7af69ddee4198444241bd5d0bd085eeee05380a7