Introduces an EditingDomain to provide access to a ResourceSet
This commit adds an EditingDomain to org.eclipse.capra.core. It can be
accessed via the static methods in EditingDomainHelper and provides
access to a unified resource set that is now used throughout Eclipse
Capra to access the trace and artifact models. We thus solve a number of
issues that were present in previous versions of Eclipse Capra:
 * We avoid constant loading and reloading of the models from disk.
 * We avoid that multiple instances of the trace model are present in
different resource sets at the same time.
 * We avoid loading the trace model into resource sets from other
editors or workbench parts.
Overall, this should provide performance and stability improvements
throughout an Eclipse installation that uses Capra.

Future commits will also deprecate some of the current API: since the
new way to handle resource sets does not require explicit handling, we
will deprecate API that uses ResourceSets as a parameter, e.g., for
loading trace models.
Change-Id: I56a689c49d1c9ec2ebed1e35adccebba7fbe9b66