UI improvements in OfficeView and fixed potential NPE
The office view now only shows the menu items for selecting a sheet and
removing all entries if an Excel sheet is actually loaded into the view.
In addition, an additional check was introduced to ensure that sheets
are only accessed when a file is loaded into the view to avoid a
potential NPE when calling OfficeView.isSheetEmptyMap().