Added ability to lock current diagram to PlantUML view
This commit adds a new entry in the Capra PlantUML View context menu
that allows to lock the current diagram, i.e., preventing the diagram
from updating if the selection changes. This can be helpful, e.g., when
working with change impact analysis where the graph should remain
visible when opening other artifacts.