[Fix] DefaultPriorityHandler now takes the artifact type into account
This commit fixes a long-standing bug that caused Eclipse Capra to select
the incorrect ArtifactHandler in certain situations. Essentially, up until
now, the DefaultPriorityHandler created a hierarchy of ArtifactHandlers
and always selected the one with the highest priority, independent of the
concrete artifact that was handled. This could lead to situations where
an incorrect ArtifactHandler was selected. With the new commit, the
DefaultPriorityHandler only returns an ArtifactHandler that fits the
artifact. The test cases test for this now.

Change-Id: I6a422552a7dd2ac650a603cc7134efa639a2a11d