[Fix] Trace detection no longer dependent on object ref
The GenericTraceMetaModelAdapter for the standard trace meta model
that Eclipse Capra ships with detected the presence of a link between
two artifacts by using contains() which falls back on the objects'
equals() implementation. This could cause problems if the underlying
models were reloaded in a different ResourceSet. This commit now
changes this behaviour and compares the elements using
EMFHelper.isElementInList() which uses the identifiers of the
artifacts instead.

Change-Id: Ia83dd71d0329cc8bf4ffd2e09574d61620f37bae