[Releng] Switched to pomless Tycho
With this commit, Eclipse Capra switches to a pomless Tycho build.
This has the advantage that we do not need to maintain individual
pom.xml files any more and that Tycho picks up all relevant assets
by itself. This means less places to update version numbers, etc.
It also enforces good practice w.r.t. naming plugins in properties
files. Some issues with this are already fixed in this commit.

Change-Id: Ifecb320babaf07c43385e2f9d43ddab7648506c0