Switched build to Java 11 and Eclipse 4.10
Eclipse Capra used to be based on Jave 8 and Eclipse Neon. This
backward-compatibility is starting to show its age: we cannot use
currect Tycho versions and are running into compatibility issues.
This commit thus switches the build to Java 11 and replaces the
standard target platform used by Tycho with one for Eclipse 4.10.
This ensures that we can make use of current features and new
developments in the releng environment.

This commit contains one not-so-nice hack: the FeatureIDE tests
now depend explicitly on javax.annotation. That package has been
deprecated in Java 9 and is therefore no longer part of the JRE.
All other tests pass, but the FeatureIDE tests fail to start since
Tycho does not explicitly add the package to the config.ini. It is
not clear why, but making the dependency explicit resolves the issue
for now.

Change-Id: Id39fd4c6c923fa3650400f198480d135c3905f41