[Fix] Tooltips in trace matrix now shown reliably
When scrolling in the trace matrix view, tooltips were not always
shown reliably since the row and column for the entry was calculated
relative to the currently displayed part of the matrix. This commit
fixes this by using the absolute position of the cell in the matrix
to find the underlying connection. Thus, tooltips in both the header
row as well as in the matrix body are now shown correctly.

Change-Id: I1140aa51073b107de1a141e17c0385721b9bcb1a