Address Sonar issues in Office handling
This commit addresses several issues in the code for handling
Office and Google documents reported by Sonar. It also fixes a
bug that prevented the "Open File" dialog from showing if no
Office document was previously loaded via drag & drop.
In addition, the menu items in the menu to select a sheet
in an Excel file now behave correctly.

Change-Id: I5169ffbb14101e0082ea380baa72c0b03976b28f