Updated API and API documentation
The recent changes in the API made an update necessary. At the same
time, the documentation needed to be updated to make the contracts
for the different methods in key interfaces clearer.
This patch addresses both issues by removing superfluous methods,
simplify method signatures, and providing clearer definitions of the
method contracts that reflect the API evolution.

Change-Id: Icf99a4ca4784e39376c4be65b770133d9d413957