Added UI support for directed trace links
This commit updated the trace creation view with support for
directed trace links. For this purpose, the list of artifacts
now contains check boxes for each artifact. If the checkbox is
selected, the artifact is regarded as a source, otherwise as
a target. The commit also adds a check to determine if at least
one source has been selected. The first artifact added to the
trace creation view is automatically selected as a source.

Change-Id: Iab8c831160f5873d674be9391110fee91e0f5307