Changed assertions in test cases
Our test cases used a lot of assertTrue() and assertFalse() calls
to check if two instances were equal. This is not optimal since
JUnit does not provide a good warning message if the assertion
fails. This change switches these statements to assertEquals() or
assertNotEquals() respectively and thus should provide better
output on regressions in the future.

Change-Id: Ie5be602ab8dbaf68568f048e9def5aebd779d98f