Added tests for handling of transitive trace links
Adding validations for methods getInternalElementsTransitive()
and getTransitivelyConnectedElements() for different types of
connections and parameter types in class AbstractMetaModel and
GenericMetaModelAdapter. The first scenario testGenericMetaModelForConnectedElements()
checks if the transitive elements are identified by getTransitivelyConnectedElements()
method when present and that traceModels without transitive links return just the direct connections.
The second test scenario testAbstractMetaModelForTransitiveElements() validates that getInternalElementsTransitive()
is also able to identify transitive elements for objects or tracelinks.

Signed-off-by: Mihaela Grubii <laborator2test@gmail.com>
Change-Id: Ieb4df3937f0ba7be268a7bf96015c60c93c75edc