Test to cover attributes vlidation in EMFHelper
Validating the getIdentifier() method in EMFHelper for
valid attribute value and exceptional values, comparing the identifiers
of similar objects. Testing uncovered validations for
tryGetAnyAttribute() covering valid and invalid attributes list as
well as tryGetNameAttribute() and getNameAttribute() for
objects that have attribute name as a null and the result after setting name as valid.
Validating those methods for a list of attributes, one element in list and
an empty list.

Signed-off-by: Mihaela Grubii <laborator2test@gmail.com>
Change-Id: Ied486335326738d350b473c7d0ccbdbae084b142