[Fix] IndexOutOfBoundsException in ArtifactHelper
Addresses an issue where an exception was thrown if the identifier of an EObject did not contain
a colon. That can happen if no name or no attribute for an artifact wrapper could be found and
only the classname is returned. The situation is rare and should not occur in practice, but it
can with a custom traceability model that is not quite conformant.

Change-Id: I808f163b5861fc51d7a5adaa19f516b74035e047