Test artifacts & wrapper logic for ArtifactHelper
Adding to class TestEMFArtifactHelper validations for Artifact Helper
createWrapper() and createwrappers() and validating special use cases
for valid and invalid input parameters, and comparing if createWrapper()
and unwrapWrapper() methods are consistent and reversible. Covering
getArtifactLabel() for parameters with no handlers, as well as
EObject parameters such as wrapper, RelatedTo, artifactModel. Validating
getArtifactLocation() for different valid and invalid  parameter types
and validating the output URI on containing the expected response
structure. Adding a conditional that checks presence of a special
character in ArtifactHelper, so it protects from an error and helps
returning message for such elements.

Signed-off-by: Mihaela Grubii <laborator2test@gmail.com>
Change-Id: I7b3521bdc650ddaf49391e1abff535dc351bd7ae