Changed naming scheme for Excel spreadsheets
Previously, the display name for Office documents used the same
representation as in the Office selection view. That meant that
depending on the preferences, large swaths of text were displayed
in the traceability graph and stored in the artifact wrapper.
This commit simplifies that and now uses the filename, the sheet
and the ID to construct a display name for Excel spreadsheets,
leading to a much more compact representation.

Change-Id: I96d7d85d7bf4e4263df6750013508d9be4993d56