Added ability to remove links to matrix view
So far, Eclipse Capra has lacked a way to delete trace links from
the user interface. This change adds this ability to the matrix view.
The selected link in the matrix view can now be deleted with a click
on the corresponding toolbar button.

Caution: right now, this action is not undoable.
Change-Id: Icc40b5eb5de6ccb2ef8fae8ba47ad9f608fe138b