PlantUML view now shows links vertically
This commit changes the layout of the graphs PlantUML renders from a
horizontal layout to a vertical one. This improves readability, in
particular in cases where an element is linked to a lot of others.
Previously, the diagram was cut off after a certain number of pixels
and therefore not fully visible. It was also difficult to navigate
the diagram.

Change-Id: I54f28590fcc7fb0c0a32de0c7c2734aadbf4e310