[Fix] getConnectedElements() set incorrect origin
If getConnectedElements() was called with a trace link as the argument,
the origin of the resulting Connection instance was sometimes set
incorrectly. This could lead to erroneous visualisations, but also
cause trouble when processing the trace links.
This change addresses this fix and also adds an additional test case
for this situation.

Change-Id: Idec2f423141f7a2d1daa949547d54312974f1b69