Project that stores traces is now configurable
This patch adds the ability to configure the project in which
the trace links are stored. If the project name is changed in
the preferences, the trace and artifact models are written to
the new project. The old project remains in the workspace.

To achieve this, this patch also adds a preference text editor
that uses a regex to check if an entry is correct.

Signed-off-by: Mihaela Grubii <laborator2test@gmail.com>
Change-Id: I571c29f434bbde8577647c9d1006b2cb65ee4340