Added operation to update traces
The new UpdateTraceOperation updates structural features of trace links
which are identified with their name to a given value. Internally, the
operation executes a write transaction on the provided trace link. The
current implementation does not automatically update the wrapping
connection if the structural feature that is being changed is the
origins or the targets.

Change-Id: Ifc40b8aa05e9f098abe88fd8bf97e95bbcbd77ca