Added bundle for trace and artifact metadata
This patch is the first in a row of patches that will add metadata
handling to Eclipse Capra. It introduces an additional model that
stores the metadata as well as a simple metadata model that will
become Capra's default.

The metadata model will also be configurable via the extension
point mechanism. This means that it can be switched out for a
different one without influencing the trace model or the artifact
model.

Change-Id: Ia5ce08751352d7ba7cc2f51987123dae8da6cdd3